Create Python script for a Raspberry Pi to bridge messages between a CAN bus and an MQTT broker
Here’s a Python script for a Raspberry Pi to bridge messages between a CAN bus and an MQTT broker. This script will read messages from the CAN bus and publish them to an MQTT topic, and it will subscribe to a topic to receive messages to be sent back to the CAN bus.